requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile -P pyzmq
#
pyzmq==18.0.1
six==1.12.0
